refactor: extract aggregation controller
